add conversation + conversationMessage

+++ schemas/favorites.ts
@@ -2,8 +2,8 @@ import z from 'zod'
 import { IFavorite } from '../types/favorite.ts'
 
 export const favoriteSchema = z.object({
-    user_id: z.number().positive(),
-    model_id: z.number().positive(),
+    user_id: z.number().positive().int(),
+    model_id: z.number().positive().int(),
 })
 
 export const validFavoriteData = (favoriteData: IFavorite) => {
